requirements.txt
# Core dependencies
Pillow>=10.0.0
piexif>=1.1.3
reverse-geocoder>=1.5.1
pycountry>=22.3.5

# AI tagging
openai>=1.0.0

# Optional: HEIC/HEIF support
pillow-heif>=0.13.0

# Database (Phase 1)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# CLI utilities (Phase 2)
tqdm>=4.65.0

# Future web interface (Phase 3)
# flask>=3.0.0
# flask-cors>=4.0.0
